added login required to prevent the users that are not super user from adding, editting and deleting products in the store

+++ products/views.py
@@ -4,6 +4,7 @@ from django.db.models.functions import Lower
 from .models import Product, Category
 from .forms import ProductForm
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 
 # Create your views here.
@@ -67,10 +68,15 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 # view for adding a product
+@login_required
 def add_product(request):
     """
     This div will be used to add product to the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the store owner can do it.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -91,8 +97,14 @@ def add_product(request):
     return render(request, template, context)
 
 # the view to edit a product
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the store owner can do it.')
+        return redirect(reverse('home'))
+
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -114,8 +126,15 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+# deleting functionality for superuser
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the store owner can do it.')
+        return redirect(reverse('home'))
+
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product is deleted!')
